feat(uncrustify): replace wildcards with directory names processing
Wildcards in '--file-name' and '--file-list' parameter values are no longer supported.
Instead these parameters now accept directory names as well as files.
If a directory is passed, all files in it will be added recursively.
Also added support for multiple values of '--pattern-form' and '--file-list' parameters.
This change was made to match other static analysers behaviour.

Change-Id: Ib640d81e33360385ccc0333b3ea9a33ec6ec1b5d