fix(ci): config didn't verify dependencies correctly
The change fixes the issue with importing modules from the host and
the issue with performing testing using outdated versions of
dependencies.

Using of "--system-site-packages" command-line parameter when creating
venv allows importing host system modules from inside venv. If some
dependency is not specified in the setup.py, but is installed on host,
with this option it will be successfully imported during the test.
However, it will not be installed during normal installation of the
Universum in the production environment. Therefore, this option leads
to false positive test result and must not be used during testing.

The caching provided by "--system-site-packages" is not needed for
testing, because venv uses shared cache with the host system
automatically.

Considering the above, it makes sense to remove
"--system-site-packages" command-line parameter.

Also, if "pip" is launched without "-U" command-line parameter, it
doesn't update dependencies. This leads to a potential issue, when
some dependency updated and did break the test, but test still
succeeded, because it continued to use old version of the dependency.
Because of these reasons, added "-U" to detect compatibility issues as
soon as dependency updates.