fix(test): nonci test started to fail due to 'sh'
The reason of the failure is that 'sh' changed the format of the
command line string returned for the launched process, and the
test is designed to check exactly one format.

Fixed the failure by avoiding checking of the command line
presence in logs at all.

Also, improved the test to check more features of the nonci mode.