fix(test): perforce container starts slowly
perforce has changed the structure of its log directories and 'tail'
command produces the error message in an endless loop. Because of
that, if the container is already started and tests attempt to reuse
it, it takes too much time to parse an enormous log. This happens
when running tests from the PyCharm.

Fixed by using correct path to logs file. Also, lowered the verbosity
level to 1. The default level is 3 and it produces too much logs.