fix(swarm): tests status for non-latest revisions
The fix disables reporting of the tests status to Swarm for intermediate
review revisions. Test status is shown as a red or green block on a
review page, but it looks like global status of the entire review and
has no visual relation to the review revision. Reporting the status of
non-latest review revisions can give the wrong impression to the viewer
of the review.