fix(gravity): incorrect type in Dependency
The Dependency class was annotated as using the same type for both
source and and target module of the dependency. Because of this
incorrect annotation, mypy produces the error at the place in the code,
where the source module calls the factory to construct the target
module.

Fixed by using two distinct type variables: one for source and one for
target.

Also, added type annotation in one method in submitter to properly
demonstrate the issue.