feat(args): printing multiple errors in p4, submit (#495)
Also removed tests with deleting fields from settings objects. If the
field is absent, the error state mechanism is able to detect it and
remember the error. However, the code in the module still can access the
field. Such accesses result in exceptions raised from gravity.

This is not an issue for normal operation, because when user launches
the script, all settings fields are present.