fix(test): set artifacts dir for nonci tests
Before this fix the artifacts directory was not specified for nonci
tests running in docker. This resulted in universum automatically
calculating it as ‘artifacts’ subdirectory of the current directory,
which is root (“/”). This is confusing and also results in artifacts
missing in a proper place.

Fixed by always specifying artifacts directory for running universum in
docker by using ‘-ad’ command-line parameter. Also fixed artifact checks
in tests accordingly.