feat(args): printing multiple errors in swarm, vcs
Fixed the check for error state in 'MainVcs’ module. If the check is implemented
for error state, as it was before, the code doesn’t try to construct the code
review module in case of any errors across the entire universum. As the result,
there is no check for errors in the parameters of the code review modules and
these errors are not reported to user.

Moved the check for code review and producing error from VCS drivers to the
'MainVcs’ module.

Fixed minor issue with unnecessary try..except block in 'Vcs’ module. In the
past the attempt to access missing vcs type parameter in module settings would
result in AttributeError exception. There were no other sources of this
exception in the code block. However, since there is an explicit check for the
vcs type setting, the try..except block is not needed anymore.