fix(analyzer): wrong error on missing input file
When the passed input file is missing, there were no explicit error
message. Instead, the empty list of files to analyze was passed to the
actual analysis tool. As the result, the error message was incorrect and
hard to understand. Fixed by producing warning on each missing input
file and producing error if the final list is empty.