fix(analyzer): wrong error on missing input file
When the passed input file is missing, there were no explicit error
message. Instead, the empty list of files to analyze was passed to the
actual analysis tool. As the result, the error message was incorrect and
hard to understand. Fixed by producing warning on each missing input
file and producing error if the final list is empty.

Also removed potential duplication of files passed to the actual
analysis tool by switching from list to set. Duplication may happen if
user passes several masks as "--files" argument and these masks expand
to intersecting sets of files.